vrynt-ml-internship-gallery-style/requirements.txt
boto3
streamlit
requests
numpy
soundfile
pydub

# diffusion demo
diffusers
transformers
accelerate
safetensors
Pillow

# notebook eval
jiwer
nbformat

# style transfer
torch
torchvision
